@@ -1,15 +1,12 @@
-'use strict';
-var test = require('ava');
-var stringIncludes = require('./');
+import test from 'ava';
+import m from './';
 
-test(function (t) {
-	var str = 'To be, or not to be, that is the question';
+test(t => {
+	const str = 'To be, or not to be, that is the question';
 
-	t.assert(stringIncludes(str, 'To be'));
-	t.assert(stringIncludes(str, 'question'));
-	t.assert(!stringIncludes(str, 'nonexistent'));
-	t.assert(!stringIncludes(str, 'To be', 1));
-	t.assert(!stringIncludes(str, 'TO BE'));
-
-	t.end();
+	t.true(m(str, 'To be'));
+	t.true(m(str, 'question'));
+	t.false(m(str, 'nonexistent'));
+	t.false(m(str, 'To be', 1));
+	t.false(m(str, 'TO BE'));
 });
